+++ scvi/dataloaders/__init__.py
@@ -1,6 +1,7 @@
 from ._ann_dataloader import AnnDataLoader
 from ._anntorchdataset import AnnTorchDataset
 from ._concat_dataloader import ConcatDataLoader
+from ._data_splitting import DataSplitter, SemiSupervisedDataSplitter
 from ._semi_dataloader import SemiSupervisedDataLoader
 
 __all__ = [
@@ -8,4 +9,6 @@ __all__ = [
     "AnnTorchDataset",
     "ConcatDataLoader",
     "SemiSupervisedDataLoader",
+    "DataSplitter",
+    "SemiSupervisedDataSplitter",
 ]
