+++ scvi/model/_scanvi.py
@@ -9,9 +9,13 @@ from pandas.api.types import CategoricalDtype
 
 from scvi import _CONSTANTS  # , settings
 from scvi._compat import Literal
-from scvi.compose._training import SemiSupervisedDataSplitter, TrainRunner
+from scvi.compose._training import TrainRunner
 from scvi.data._anndata import _make_obs_column_categorical
-from scvi.dataloaders import AnnDataLoader, SemiSupervisedDataLoader
+from scvi.dataloaders import (
+    AnnDataLoader,
+    SemiSupervisedDataLoader,
+    SemiSupervisedDataSplitter,
+)
 from scvi.lightning import SemiSupervisedTrainingPlan
 from scvi.lightning._callbacks import SubSampleLabels
 from scvi.model._scvi import SCVI
