+++ scvi/model/_totalvi.py
@@ -12,10 +12,10 @@ from scvi import _CONSTANTS
 from scvi._compat import Literal
 from scvi._docs import doc_differential_expression
 from scvi._utils import _doc_params
-from scvi.compose._training import DataSplitter, TrainRunner
+from scvi.compose._training import TrainRunner
 from scvi.data import get_from_registry
 from scvi.data._utils import _check_nonnegative_integers
-from scvi.dataloaders import AnnDataLoader
+from scvi.dataloaders import AnnDataLoader, DataSplitter
 from scvi.lightning import AdversarialTrainingPlan
 from scvi.model._utils import (
     _get_batch_code_from_category,
